requirements.txt
nfcpy
pandas